Improved performance of toggling combat ui indicators
